backend/requirements.txt
Flask
openai
python-dotenv
gunicorn
flask-cors
agentops
